newlinkedien/requirements.txt
beautifulsoup4==4.9.1
bs4==0.0.1
joblib==0.16.0
mysql-connector==2.2.9
numpy==1.19.1
pandas==1.1.0
python-dateutil==2.8.1
pytz==2020.1
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
urllib3==1.25.10
